+++ packages/snap-controller/src/Autocomplete/AutocompleteController.ts
@@ -170,7 +170,12 @@ export class AutocompleteController extends AbstractController {
 			}
 		}
 
-		inputElement?.dispatchEvent(new Event('keyup'));
+		//auto select first trending term?
+		if (!this.store.state?.input && this.store.trending?.length && !this.store.terms?.length && this.config.settings?.trending?.showResults) {
+			this.store.trending[0].preview();
+		} else {
+			inputElement?.dispatchEvent(new Event('keyup'));
+		}
 	}
 
 	reset(): void {