+++ CHANGELOG.md
@@ -3,6 +3,24 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.22.0](https://github.com/searchspring/snap/compare/v0.21.1...v0.22.0) (2022-03-09)
+
+
+### Bug Fixes
+
+* **facetstore:** altering logic to ensure meta and facet details are in agreement ([efd9ceb](https://github.com/searchspring/snap/commit/efd9cebee93ea65b9b218160d23303f0215ab573))
+
+
+### Features
+
+* **autocomplete:** ac showResults config prop for showing trending term results ([7a1d71e](https://github.com/searchspring/snap/commit/7a1d71e2f15c2beff4c9403cb348a017d41f64a3))
+* **client:** adding 'facets.autoDrillDown' request parameter to transformer ([1726235](https://github.com/searchspring/snap/commit/1726235c56ceef0b3bdaa28efe0a0e1fa5229241))
+* **findercontroller:** adding default parameters to use 'facets.autoDrillDown' by default ([fa2df7e](https://github.com/searchspring/snap/commit/fa2df7e2639abc3d087b399c80f757a6455e9dbe))
+
+
+
+
+
 ## [0.21.1](https://github.com/searchspring/snap/compare/v0.21.0...v0.21.1) (2022-03-01)