+++ packages/snap-client/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.22.0](https://github.com/searchspring/snap/compare/v0.21.1...v0.22.0) (2022-03-09)
+
+
+### Features
+
+* **client:** adding 'facets.autoDrillDown' request parameter to transformer ([1726235](https://github.com/searchspring/snap/commit/1726235c56ceef0b3bdaa28efe0a0e1fa5229241))
+
+
+
+
+
 ## [0.21.1](https://github.com/searchspring/snap/compare/v0.21.0...v0.21.1) (2022-03-01)
 
 **Note:** Version bump only for package @searchspring/snap-client